further refactored namespace separatation and copied some sqlite and mysql methods to mysql and sqlite classes
