+++ src/lib.rs
@@ -9,10 +9,8 @@ pub mod models {
     pub mod table_data;
 }
 
-pub mod data_comparer;
 pub mod data_creator;
 pub mod data_exporter;
 pub mod data_querier;
 pub mod log;
 pub mod processor;
-pub mod ui;
